Refactored import of classes, functions and variables. Now need use direct import.
