+++ src/app/__init__.py
@@ -1,14 +1,7 @@
-from .base import Base, Session
-from .config import config
-from .migrate import (create_db, create_tables, create_user, insert_data_in_db,
-                      drop_tables)
-from .models import User, Password, SessionObject
-from .resources import Home, Smoke, EntityResource, UserResource, \
-    UserListResource, PasswordListResource, PasswordResource
-
 from flask import Flask
 from flask_restful import Api
 
+from .config import config
 
 app = Flask(__name__)
 app.config.from_object(config['default'])
